Remove Tensors from LoopNest.
