+++ torch/csrc/jit/tensorexpr/analysis.cpp
@@ -1,0 +1,72 @@
+#include <torch/csrc/jit/tensorexpr/analysis.h>
+
+#include <queue>
+
+namespace torch {
+namespace jit {
+namespace tensorexpr {
+
+class DepTracker : public IRVisitor {
+ public:
+  std::vector<Tensor*> findUsedTensors(Tensor* tensor) {
+    used_tensors.clear();
+    tensor->stmt()->accept(this);
+    return used_tensors;
+  }
+
+ private:
+  void visit(const FunctionCall* v) override {
+    used_tensors.push_back(const_cast<Tensor*>(v->tensor())); // NOLINT
+  }
+
+  std::vector<Tensor*> used_tensors;
+};
+
+std::vector<Tensor*> findAllNeededTensors(const std::vector<Tensor*>& tensors) {
+  DepTracker d;
+  std::queue<Tensor*> q;
+  std::unordered_set<Tensor*> queued;
+  std::vector<Tensor*> result;
+  std::unordered_set<Tensor*> processed;
+  for (Tensor* t : tensors) {
+    if (queued.insert(t).second) {
+      q.push(t);
+    }
+  }
+  while (!q.empty()) {
+    Tensor* t = q.front();
+    q.pop();
+    queued.erase(t);
+    std::vector<Tensor*> deps = d.findUsedTensors(t);
+    bool all_processed = true;
+    for (Tensor* dep : deps) {
+      if (!processed.count(dep)) {
+        if (queued.insert(dep).second) {
+          q.push(dep);
+        }
+        all_processed = false;
+      }
+    }
+    if (all_processed) {
+      result.push_back(t);
+      if (processed.count(t)) {
+        throw malformed_input("failure to find all processed Tensors");
+      }
+
+      processed.insert(t);
+    } else {
+      if (queued.count(t)) {
+        throw malformed_input("failure to find all queued Tensors");
+      }
+
+      q.push(t);
+      queued.insert(t);
+    }
+  }
+
+  return result;
+}
+
+} // namespace tensorexpr
+} // namespace jit
+} // namespace torch
