+++ torch/csrc/jit/tensorexpr/analysis.h
@@ -189,6 +189,9 @@ class CreateBufferMap : public IRVisitor {
   }
   std::unordered_map<std::string, const Buf*> map_input_to_tensor_bufs_;
 };
+
+std::vector<Tensor*> findAllNeededTensors(const std::vector<Tensor*>& tensors);
+
 } // namespace tensorexpr
 } // namespace jit
 } // namespace torch
