+++ torch/csrc/jit/tensorexpr/loopnest.h
@@ -26,6 +26,9 @@ class TORCH_API LoopNest {
  public:
   // A constructor for building a LoopNest from a list of Tensors
   LoopNest(const std::vector<Tensor*>& output_tensors);
+  LoopNest(
+      const std::vector<Tensor*>& output_tensors,
+      const std::vector<Tensor*>& tensors_to_compute);
 
   // A constructor for building a LoopNest from a pre-baked Stmt and meta-info
   // TODO: Nuke intermediate_bufs_ from here if they can be deduced.
@@ -122,8 +125,9 @@ class TORCH_API LoopNest {
   }
 
  private:
-  std::vector<Tensor*> findAllNeededTensors(
-      const std::vector<Tensor*>& tensors);
+  void initialize(
+      const std::vector<Tensor*>& output_tensors,
+      const std::vector<Tensor*>& tensors_to_compute);
   Stmt* insertAllocFree(Stmt* stmt);
 
   Stmt* root_stmt_;
